+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   cast it to a string.
 * Adds node type iterators to the container class (e.g. `eachComment`).
 * Adds filter function to the container class.
+* Adds split function to the container class.
 
 ---